+++ packages/ember-easyForm/lib/helpers/form-for.js
@@ -1,4 +1,4 @@
 Ember.Handlebars.registerHelper('form-for', function(object, options) {
-  options.hash.contentBinding = object;
+  options.data.keywords.formForModelPath = object;
   return Ember.Handlebars.helpers.view.call(this, Ember.EasyForm.Form, options);
 });