+++ packages/ember-easyForm/lib/views/base_view.js
@@ -10,5 +10,14 @@ Ember.EasyForm.BaseView = Ember.View.extend({
     } else {
       return 'default';
     }
-  })
+  }),
+  formForModel: function(){
+    var formForModelPath = this.get('templateData.keywords.formForModelPath');
+
+    if (formForModelPath) {
+      return this.get('context.' + formForModelPath);
+    } else {
+      return this.get('context');
+    }
+  }.property(),
 });